bug fix in bimodal output
