+++ allennlp/modules/transformer/__init__.py
@@ -8,7 +8,12 @@ from allennlp.modules.transformer.attention_scores import (
 from allennlp.modules.transformer.positional_encoding import SinusoidalPositionalEncoding
 
 from allennlp.modules.transformer.transformer_module import TransformerModule
-from allennlp.modules.transformer.transformer_embeddings import Embeddings, TransformerEmbeddings
+from allennlp.modules.transformer.transformer_embeddings import (
+    Embeddings,
+    TextEmbeddings,
+    TransformerEmbeddings,
+    ImageFeatureEmbeddings,
+)
 from allennlp.modules.transformer.self_attention import SelfAttention
 from allennlp.modules.transformer.activation_layer import ActivationLayer
 from allennlp.modules.transformer.transformer_layer import AttentionLayer, TransformerLayer
