+++ test_fixtures/vilbert/experiment.json
@@ -22,23 +22,23 @@
       "hidden_dim": 28
     },
     "encoder": {
-      "text_num_hidden_layers": 2,
-      "image_num_hidden_layers": 2,
-      "text_hidden_size": 22,
-      "image_hidden_size": 28,
+      "num_hidden_layers1": 2,
+      "num_hidden_layers2": 2,
+      "hidden_size1": 22,
+      "hidden_size2": 28,
       "combined_hidden_size": 26,
-      "text_intermediate_size": 5,
-      "image_intermediate_size": 7,
+      "intermediate_size1": 5,
+      "intermediate_size2": 7,
       "num_attention_heads": 2,
-      "text_attention_dropout": 0.0,
-      "image_attention_dropout": 0.0,
-      "text_hidden_dropout": 0.0,
-      "image_hidden_dropout": 0.0,
+      "attention_dropout1": 0.0,
+      "attention_dropout2": 0.0,
+      "hidden_dropout1": 0.0,
+      "hidden_dropout2": 0.0,
       "activation": "gelu",
-      "v_biattention_id": [0, 1, 2],
-      "t_biattention_id": [0, 1, 2],
-      "fixed_t_layer": 0,
-      "fixed_v_layer": 0
+      "biattention_id1": [0, 1, 2],
+      "biattention_id2": [0, 1, 2],
+      "fixed_layer1": 0,
+      "fixed_layer2": 0
     },
     "pooled_output_dim": 3,
     "fusion_method": "sum"
